o Keep copies of already delivered replies for 20 seconds and retransmit if the request with matching cookie comes in;
o When request within existing session comes in match it not
only against active nodes, but also against "pending" too. This
prevents storm of useless broadcasts when node drop our of the
sight when it has lot of session bound to it, which may cause
domino affect clogging up command pipe and taking the whole
thing down temporarily.